+++ src/_nanostores/index.js
@@ -1,0 +1,2 @@
+export * from 'nanostores'
+export { atom as _atom_, computed as _computed_ } from 'nanostores'