+++ src/atom_/index.js
@@ -1,5 +1,5 @@
-import { atom as _atom_ } from 'nanostores'
-import { writable_fn_ } from '../_private/index.js'
+import { _atom_ } from '../_nanostores/index.js'
+import { writable_fn_ } from '../writable_fn_/index.js'
 export const atom_ = initialValue=>{
 	return writable_fn_(_atom_(initialValue))
 }