+++ src/computed_/index.js
@@ -1,5 +1,5 @@
 import { computed as _computed_ } from 'nanostores'
-import { readable_fn_ } from '../_private/index.js'
+import { readable_fn_ } from '../readable_fn_/index.js'
 export const computed_ = (stores, cb)=>{
 	return readable_fn_(_computed_(stores, cb))
 }