+++ src/writable_fn_/index.d.ts
@@ -1,0 +1,3 @@
+import { ReadableAtom } from 'nanostores'
+import { WritableAtom_ } from '../WritableAtom_'
+export declare function writable_fn_<Value>(atom:ReadableAtom):WritableAtom_<Value>