+++ src/writable_fn_/index.js
@@ -1,17 +1,3 @@
-export function readable_fn_(atom, rest = {}) {
-	function fn() {
-		return atom.get()
-	}
-	return new Proxy(fn, {
-		get(target, prop, receiver) {
-			if (prop === '_' || prop === '$') {
-				return atom.get()
-			}
-			return Reflect.get(atom, prop, atom)
-		},
-		...rest
-	})
-}
 export function writable_fn_(atom) {
 	function fn(_) {
 		return arguments.length ? atom.set(_) : atom.get()